Bug 1346211 - Part 4: Split functions, create additional abstractions, and cleanup in preparation for ICU time zone changes. r=Waldo
UltraBlame original commit: fe6637fcb77005211e23c8adb38ac32dbbbb4414

+++ js/src/vm/Time.h
@@ -57,7 +57,7 @@ PRMJ_NowShutdown() {}
 
 
 extern size_t
-PRMJ_FormatTime(char* buf, int buflen, const char* fmt, const PRMJTime* tm,
+PRMJ_FormatTime(char* buf, size_t buflen, const char* fmt, const PRMJTime* tm,
                 int timeZoneYear, int offsetInSeconds);
